% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.collapse_pop}
\alias{.collapse_pop}
\title{Get a searchable version of the marker levels}
\usage{
.collapse_pop(pop, search = FALSE)
}
\arguments{
\item{search}{logical. If \code{TRUE}, then square brackets are concatenated
around each of the elements in the level strings.}
}
\description{
Get a searchable version of the marker levels
}
\examples{
.collapse_pop(c("CD4" = "-", CD3" = "-"), search = FALSE); .collapse_pop(c("CD4" = "-", CD3" = "-"), search = TRUE)
}
