% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_write.R
\name{.faust_fcs_write}
\alias{.faust_fcs_write}
\title{Save FAUST subset as an FCS file}
\usage{
.faust_fcs_write(
  project_path,
  fr_source,
  sample_name,
  sel_sample,
  pop,
  dir_save,
  trans_fn = NULL,
  trans_chnl = NULL
)
}
\arguments{
\item{project_path}{character. FAUST project directory.}

\item{fr_source}{GatingSet or character vector. If a \code{GatingSet}, then
the flowFrames within the GatingSet are used to create the output fcs files.
If a character vector,
then must specify a directory containing FCS files. These FCS
files are used to create the output fcs files.}

\item{sample_name}{\code{character vector}.
Specifies names of samples in order
in which they are found in \code{gs}.}

\item{sel_sample}{\code{character vector}. Character vector specifying
the names of the samples
(as saved by FAUST as folder names in the analysis map
and equivalent to \code{gs[[i]]@name}
where i is an index in \code{gs}.) that
are to have their FAUST pops saved.}

\item{pop}{\code{list}, \code{named character vector}
or unnamed character vector of length one.
If an unnamed character vector,
then all cells matching that full FAUST annotation will be returned
(e.g. "CD3+CD8+CD45RA+IFNg+IL2-TNF+" or, equivalently (depending
on the number of sub-populations FAUST detects)
"CD3~2~2~CD8~2~2~CD45RA~2~2~IFNg~2~2~IL2~1~2~TNF~2~2~")).
If a named \code{character vector},
then all cells matching the set of marker levels are returned.
(names are markers, and elements are levels, e.g. c("CD3" = 2, "CD8" = 2)).
If a \code{list}, then each
element must be a named \code{character vector}, and then
all cells matching either of
these specified sets of marker levels are returned.}

\item{dir_save}{\code{character}.
Specifies directory in which files are saved.}

\item{trans_fn}{function. If supplied, this function
is applied to the expression data.
Useful for back-transformation.
If \code{NULL}, then no transformation is applied. Default is \code{NULL}.}

\item{trans_chnl}{character vector.
If specified, \code{trans_fn} is applied to only these channels.
If \code{NULL} and if \code{trans_fn} is not \code{NULL},
then \code{trans_fn} is applied to entire
expression matrix. Default is \code{NULL}.}
}
\value{
\code{invisible(TRUE)}.
}
\description{
Save FAUST subset as an FCS file
}
\keyword{internal}
