% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_faust_pop.R
\name{.get_faust_pop_pop}
\alias{.get_faust_pop_pop}
\title{Return expression matrix for cells that match a single FAUST label}
\usage{
.get_faust_pop_pop(ex, faust_ann, pop)
}
\arguments{
\item{ex}{\code{numeric matrix}. Contains expression data.}

\item{faust_ann}{\code{character vector}. Specifies FAUST annotation for
each cell in \code{ex}.}

\item{pop}{\code{list} or \code{named character vector}. If a \code{character vector},
then all cells matching the set of marker levels are returned. If a \code{list}, then each
element must be a \code{character vector}, and then
all cells matching either of these specified sets of marker levels are returned.}
}
\value{
\code{numeric matrix} containing the expression levels for the markers specified.
}
\description{
Return expression matrix for cells that match a single FAUST label
}
