% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_faust_count.R
\name{.get_pop_counts}
\alias{.get_pop_counts}
\title{Get counts of subsets identified by FAUST}
\usage{
.get_pop_counts(data, pop, dem_col = c("sample", "exp_unit", "tot_count"))
}
\arguments{
\item{data}{dataframe. Dataframe where columns specify counts.}

\item{pop}{\code{named character vector} or \code{list}. If a \code{character vector},
then names specify markers and values specify levels, e.g. c("CD4" = "-", "CD8" = "+") means
that we are interested in FAUST subsets annotated "CD4-CD8+". Counts of all subsets with these
levels for these markers are plotted separately.
If a \code{list}, then each element must be a character vector as above. For a given list element,
instead of all plotting all subsets individually that have the correct level for the specified markers,
we sum over all the subsets matching the specified annotation and plot the final count.}

\item{dem_col}{\code{character vector}. Specifies names of columns in \code{data} that we wish
to keep, regardless of if they match a FAUST annotation.}
}
\value{
A dataframe with columns as specified in \code{dem_col}, \code{pop} (population name) and
\code{count} (number of cells in population).
}
\description{
Get counts of subsets identified by FAUST
}
