% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_write.R
\name{.is_faust_ann_a_match_for_marker}
\alias{.is_faust_ann_a_match_for_marker}
\title{Check if FAUST annotation has as given level for a given marker}
\usage{
.is_faust_ann_a_match_for_marker(faust_ann, marker, level)
}
\arguments{
\item{faust_ann}{\code{data.frame}. FAUST annotation table read from
<project_path>/faustData/sampleData/\if{html}{\out{<sample>}}/faustAnnotation.csv.}

\item{marker}{character. Name of marker.}

\item{level}{character. Level of marker, e.g. "1", "2" or "3".}
}
\description{
Check if FAUST annotation has as given level for a given marker
}
\keyword{internal}
