% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_faust_pop.R
\name{.save_faust_pop}
\alias{.save_faust_pop}
\title{Save FAUST subset as an FCS file}
\usage{
.save_faust_pop(
  project_path,
  fr_source,
  sample_name,
  sel_sample,
  pop,
  dir_save,
  trans_fn = NULL,
  trans_chnl = NULL
)
}
\arguments{
\item{project_path}{character. FAUST project directory.}

\item{fr_source}{GatingSet or character vector. If a \code{GatingSet}, then
the flowFrames within the GatingSet are used to create the output fcs files.
If a character vector, then must specify a directory containing FCS files. These FCS
files are used to create the output fcs files.}

\item{sample_name}{\code{character vector}. Specifies names of samples in order
in which they are found in \code{gs}.}

\item{sel_sample}{\code{character vector}. Character vector specifying
the names of the samples (as saved by FAUST as folder names in the analysis map
and equivalent to \code{gs[[i]]@name} where i is an index in \code{gs}.) that
are to have their FAUST pops saved.}

\item{pop}{\code{list} or \code{named character vector}. If a \code{character vector},
then all cells matching the set of marker levels are returned. If a \code{list}, then each
element must be a \code{character vector}, and then
all cells matching either of these specified sets of marker levels are returned.}

\item{dir_save}{\code{character}. Specifies directory in which files are saved.}

\item{trans_fn}{function. If supplied, this function is applied to the expression data.
Useful for back-transformation. If \code{NULL}, then no transformation is applied. Default is \code{NULL}.}

\item{trans_chnl}{character vector. If specified, \code{trans_fn} is applied to only these channels.
If \code{NULL} and if \code{trans_fn} is not \code{NULL}, then \code{trans_fn} is applied to entire
expression matrix. Default is \code{NULL}.}
}
\value{
\code{invisible(TRUE)}.
}
\description{
Save FAUST subset as an FCS file
}
