% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_plot_count.R
\name{faust_count_plot}
\alias{faust_count_plot}
\title{Plot counts of FAUST pops by sample}
\usage{
faust_count_plot(
  project_path,
  pop,
  font_size = 10,
  point_size_max = 2,
  breaks = NULL,
  trans_x = "asinh",
  p_width = NULL,
  p_height = NULL,
  exhaustive = FALSE,
  limitsize = FALSE
)
}
\arguments{
\item{project_path}{character. FAUST project directory.}

\item{pop}{\code{list} or \code{named character vector}.
If a \code{character vector},
then counts of all subsets matching the set of marker levels are plotted.
If a \code{list}, then each list element must be a character vector.
In that case,
one boxplot of frequencies is plotted for each element of the list of
where the frequency is
the sum of all cells that match the population specified.
If \code{NULL}, then all subsets found are plotted.}

\item{breaks}{numeric vector. If supplied, then this is the breaks
for the x-axis. If not supplied, then the default breaks are used of
`c(0, 0.001, 0.005, 0.01, 0.02, 0.5, 0.1, 0.5, 1, 2, 5, 10, 20, 50, 100)``.}

\item{trans_x}{\code{character} or \code{Transformation} object
(from \code{scales} package).
the name of a transformation object or the object itself. Built-in
transformations include "asinh", "asn", "atanh", "boxcox", "date", "exp",
"hms", "identity", "log", "log10", "log1p", "log2", "logit", "modulus",
"probability", "probit", "pseudo_log", "reciprocal", "reverse", "sqrt" and
"time". Default is \code{asinh}.}

\item{p_width}{\code{numeric}.
Width of saved plot in cm. If \code{NULL},
then set to 40.}

\item{p_height}{\code{numeric}. Height of saved plot in cm.
If \code{NULL}, then
it is set to \code{max(5, n_pop * 1.25)},
where \code{n_pop} is the number of sub-populations in \code{pop}.}

\item{exhaustive}{logical.
If \code{TRUE}, then counts are taken from the exhaustive FAUST count matrix
rather than the count matrix after excluding subsets that
don't appear in sufficiently many
experimental units. Default is \code{FALSE}.}

\item{limitsize}{logical.
If \code{TRUE}, then plots of size 125cm x 125cm or more are
not saved but an error is returned when \code{cowplot::ggsave2} is run.
Default
is \code{FALSE}.}
}
\description{
Saves plots of counts to \code{project_path/faustData/plotData/pop_stats}
for specified FAUST-identified subsets.
}
\details{
The key parameter is \code{pop}. If \code{pop} is a list,
then each element is treated as a population for which counts
are required on aggregate (rather than for individual subsets
of that population). If \code{pop} is a character vector, then
it is treated as a population for which counts are required for
individual subsets.
}
\examples{
project_path <- usethis::proj_path(ext = "/inst/extdata")
# plot all subsets of pop, as it is a character vector
pop <- c(
  "CD3" = "+", "CD4" = "+", "CD8-IgD" = "-",
  "CD20" = "-", "CD33" = "-", "CD14" = "-",
  "TCRgd-CD19" = "-"
)
plot_faust_count(
  project_path = project_path,
  pop = pop
)
# plot counts of cells matching annotation of each
# element in pop, as pop is a list
pop <- list(
  c(
    "CD3" = "+", "CD4" = "+", "CD8-IgD" = "-",
    "CD20" = "-", "CD33" = "-", "CD14" = "-",
    "TCRgd-CD19" = "-"
  ),
  c(
    "CD3" = "+", "CD4" = "-", "CD8-IgD" = "+",
    "CD20" = "-", "CD33" = "-", "CD14" = "-",
    "TCRgd-CD19" = "-"
  )
)
plot_faust_count(
  project_path = project_path,
  pop = pop
)
}
