% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{faust_get_marker}
\alias{faust_get_marker}
\title{Get the number of levels used for each marker FAUST used to cluster}
\usage{
faust_get_marker(project_path)
}
\value{
\code{Character vector}
}
\description{
Gives the markers FAUST used and the number of levels used per such marker.
Note that it only works after the SCAMP clustering has been performed, i.e.
not simply after the annotations have been approved.
}
