% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_faust_pop.R
\name{save_faust_pop}
\alias{save_faust_pop}
\title{Save FAUST subset as an FCS file}
\usage{
save_faust_pop(project_path, pop, gs, sample = NULL, trans = NULL)
}
\arguments{
\item{project_path}{character. Path to directory containing FAUST output.}

\item{pop}{\code{list} or \code{named character vector}. If a \code{character vector},
then all cells matching the set of marker levels are returned. If a \code{list}, then each
element must be a \code{character vector}, and then
all cells matching either of these specified sets of marker levels are returned.}

\item{gs}{GatingSet. Original \code{GatingSet} object to which \code{faust} was applied.}

\item{sample}{integer vector or character vector. If an integer vector, then it
specifies the indices of the samples for which to save output. If character,
then it specifies the names of the sames to save output for. If \code{NULL},
then the output for every sample is saved. Default is \code{NULL}.}

\item{trans}{function. If supplied, this function is applied to the expression data.
Useful for back-transformation, as the original
GatingSet . If \code{NULL}, then no transformation is applied. Default is \code{NULL}.
Not working at the moment.}

\item{saveFCS}{Boolean. If \code{TRUE}, then FCS files of the selected cell population(s)
and sample(s) are saved to project_path/faustData/fcsData. Default is \code{FALSE}.
population of all the selected samplesare saved to project_path/faustData/gsData. Default is \code{FALSE}.}
}
\value{
\code{invisible(TRUE)}. Side effect is the saved FCS file.
}
\description{
Save a specified FAUST-identified population to \code{project_path/faustData/fcsData}
as an FCS file for all samples gated or just a specified subset.
}
\examples{
save_faust_pop(project_path = "", pop = list("CD3" = 2),
gs = gs, sample = 1)
}
