% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plot.R
\docType{data}
\name{trans_asinh}
\alias{trans_asinh}
\title{Create inverse hyperbolic sin transformation object}
\format{
An object of class \code{transform} of length 9.
}
\usage{
trans_asinh
}
\description{
Create inverse hyperbolic sin transformation object
}
\keyword{internal}
