% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_plot.R
\name{.get_pop_match_ind}
\alias{.get_pop_match_ind}
\title{Get indices of columns that have a specified annotation}
\usage{
.get_pop_match_ind(data, pop)
}
\arguments{
\item{data}{dataframe. Columns containg FAUST-population counts.}

\item{pop}{\code{named character vector}.
Names specify marker and values specify level,
e.g. c("CD4" = "-", "CD8" = "+").}
}
\value{
\code{Integer vector}.
}
\description{
Get indices of columns that have a specified annotation
}
