% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_save.R
\name{.is_faust_ann_a_match_pop}
\alias{.is_faust_ann_a_match_pop}
\title{Check if FAUST annotation has a given level for a set of markers}
\usage{
.is_faust_ann_a_match_pop(faust_ann, pop)
}
\arguments{
\item{pop}{\code{named character vector}.
Names are marker names and values are
levels of marker.
Values must be of the form or "<num_1>" or "<num_1>~<num_2>",
where <num_1> is the level for the marker and
<num_2> is the total number of levels for
the marker.}
}
\description{
Check if FAUST annotation has a given level for a set of markers
}
\details{
This is effectively an AND statement across
all markers specified in the
single FAUST population annotation.
}
