% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_save.R
\name{.save_faust_pop}
\alias{.save_faust_pop}
\title{Save FAUST subset as an FCS file}
\usage{
.save_faust_pop(
  project_path,
  fr_source,
  sample_name,
  sel_sample,
  pop,
  dir_save,
  trans_fn = NULL,
  trans_chnl = NULL
)
}
\arguments{
\item{sample_name}{\code{character vector}. Specifies names of samples in order
in which they are found in \code{gs}.}

\item{sel_sample}{\code{character vector}. Character vector specifying
the names of the samples (as saved by FAUST as folder names in the analysis map
and equivalent to \code{gs[[i]]@name} where i is an index in \code{gs}.) that
are to have their FAUST pops saved.}

\item{dir_save}{\code{character}. Specifies directory in which files are saved.}
}
\value{
\code{invisible(TRUE)}.
}
\description{
Save FAUST subset as an FCS file
}
