% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_get_count.R, R/pop_save.R
\name{faust_fcs_write}
\alias{faust_fcs_write}
\title{Get counts of subsets identified by FAUST}
\usage{
faust_fcs_write(
  project_path,
  dir_save = NULL,
  pop,
  fr_source = NULL,
  sample = NULL,
  trans_fn = NULL,
  trans_chnl = NULL
)

faust_fcs_write(
  project_path,
  dir_save = NULL,
  pop,
  fr_source = NULL,
  sample = NULL,
  trans_fn = NULL,
  trans_chnl = NULL
)
}
\arguments{
\item{project_path}{character. FAUST project directory.}

\item{dir_save}{character.
Directory to save to.
If \code{NULL} (the default), then FCS files
are saved to
\code{file.path(project_path, "faustData", "fcsData", <pop_defn>)},
where \code{<pop_defn>} is a concatenation of the population
definition, e.g list(c("CD3" = 1, "CD4" = 2)) becomes "CD3~1~CD4~2~".
If \code{character}, then the FCS files are saved directly to this directory.
As stated before, default is \code{NULL}.}

\item{pop}{\code{list} or \code{named character vector}.
If a \code{character vector},
then all cells matching the set of marker levels are returned.
If a \code{list}, then each
element must be a \code{character vector}, and then
all cells matching either of
these specified sets of marker levels are returned.}

\item{fr_source}{GatingSet or character vector. If a \code{GatingSet}, then
the flowFrames within the GatingSet are used to create the output fcs files.
If a character vector,
then must specify a directory containing FCS files. These FCS
files are used to create the output fcs files.}

\item{sample}{integer vector or character vector.
If an integer vector, then it
specifies the indices of the samples for which to save output. If character,
then it specifies the names of the sames to save output for. If \code{NULL},
then the output for every sample is saved. Default is \code{NULL}.}

\item{trans_fn}{function. If supplied, this function
is applied to the expression data.
Useful for back-transformation.
If \code{NULL}, then no transformation is applied. Default is \code{NULL}.}

\item{trans_chnl}{character vector.
If specified, \code{trans_fn} is applied to only these channels.
If \code{NULL} and if \code{trans_fn} is not \code{NULL},
then \code{trans_fn} is applied to entire
expression matrix. Default is \code{NULL}.}

\item{data}{dataframe. Dataframe where columns specify counts. If \code{NULL},
then it is read in from \code{project_path/faustData/faustCountMatrix.rds}.}

\item{dem_col}{\code{character vector}.
Specifies names of columns in \code{data} that we wish
to keep, regardless of if they match a FAUST annotation.}

\item{exhaustive}{logical. If \code{TRUE},
then counts are taken from the exhaustive FAUST count matrix
rather than the count matrix after excluding subsets that
don't appear in sufficiently many
experimental units. Default is \code{FALSE}.}
}
\value{
A dataframe with columns as specified in
\code{dem_col}, \code{tot_count} (total number of cells classified for a
sample), \code{pop} (population name) and
\code{count} (number of cells in population).

\code{invisible(TRUE)}. Side effect is the saved FCS file.
}
\description{
Save a specified FAUST-identified population to
\code{project_path/faustData/fcsData}
as an FCS file for all samples gated or just a specified subset.
}
\examples{
# get counts for cells matching one annotation
pop <- c("CD4" = "-", "CD8" = "+")
get_pop_counts(pop = pop)
# get counts for cells matching either of the two annotations
pop <- list(c("CD4" = "-", "CD8" = "+"), c("CD8" = "-", "CD4" = "+"))
get_pop_counts(pop = pop)
faust_fcs_write(
  project_path = "", pop = list("CD3" = 2),
  gs = gs, sample = 1
)

}
